requirements.txt
# Core Deep Learning Dependencies
tensorflow>=2.8.0
keras>=2.8.0
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Web Framework
Flask>=2.0.0
flask-cors>=3.0.0

# Data Processing
albumentations>=1.1.0

# Utilities
tqdm>=4.62.0
h5py>=3.6.0